Move OrLeafPlanner to Cypher compiler 3.2
- Fix compile error in LogicalPlanProducer
